pyproject.toml
[project]
name = "ClouderaAIAgentStudio"
version = "0.0.1"
description = "Cloudera AI Agent Studio"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arize-phoenix==7.1.1",
    "arize-phoenix-otel==0.6.1",
    "crewai==0.80.0",
    "crewai-tools==0.14.0",
    "gql==3.5.0",
    "grpcio==1.65.4",
    "grpcio-tools==1.65.4",
    "litellm[proxy]==1.57.4",
    "openinference-instrumentation-crewai==0.1.3",
    "openinference-instrumentation-litellm==0.1.5",
    "opentelemetry-api==1.28.2",
    "protobuf==5.27.3",
    "psutil==6.1.0",
    "pydantic==2.10.6",
    "pysqlite3-binary==0.5.4; platform_system == 'Linux'",
    "sqlalchemy==2.0.36",
    "uv==0.5.25",
    "alembic==1.14.1",
]

[dependency-groups]
dev = [
    "autoflake==2.3.1",
    "debugpy==1.8.12",
    "pytest==8.2.2",
    "pytest-cov==5.0.0",
    "ruff==0.9.4",
]
workflow_deployment = [
    "arize-phoenix==7.1.1",
    "arize-phoenix-otel==0.6.1",
    "crewai==0.80.0",
    "gql==3.5.0",
    "openinference-instrumentation-crewai==0.1.3",
    "openinference-instrumentation-litellm==0.1.5",
    "opentelemetry-api==1.28.2",
    "pydantic==2.10.6",
    "pysqlite3-binary==0.5.4",
]

[tool.ruff]
line-length = 120
